src/dash/requirements.txt
dash
plotly
streamlit
mibian
selenium
webdriver-manager
matplotlib
pandas
scipy
gunicorn
boto3
requests
python-dateutil